Change Model to store word embeddings with subword embeddings added.
Up to this point, Model has stored word representations without
subword units added. This is a large disadvantage once we add support
for memory-mapped models, since in order to compute the embedding of a
known word, we have to touch a lot of pages.

This change updates Model and TrainModel -> Model conversions such
that Model always contains word embeddings that have their subword
representations already added.

With this change, it has become unnecessary to compute a separate
embedding matrix in Similarity. However, this is rectified in a
separate commit.