Show a progress bar while reading the vocabulary
finalfrontier followed the tradition of several other word embedding
tools to report progress in the number of tokens. However, it is
typically much more useful to know what percentage of a the corpus is
read and how long the remaining estimated time is. Therefore this change
replaces the word count by a progress bar.