Introduce BandedRangeGenerator to fix structgram NS
As @sebpuetz has pointed out, the choice of negative samples for the
structured skip-gram model is wrong. It picks an index [0, vocab_size),
however in the structured skip-gram case the output matrix consists of
context_size * 2 rows for each token. So, instead, we should first draw
a token following a zipfian distribution and then a context position
following a uniform distribution.

In this change, this is abstracted to a BandedRangeGenerator, which pick
from a 'banded' range. In this case, the bands are the tokens and the
band size is the total number of context positions (2 * context size).