Untie RNGs
We use RNGs for several purposes, such as: negative sampling and
discards. So far, we have been cloning RNGs, but this can give
problematic dependencies between the generators. E.g. I observed
biased distributions when testing BandedRangeGenerators using clones
of the same RNG for the inner generator and the uniform sampling
within a band.

This change makes it possible to use separate RNGs for negative
sampling and discards. For convenience, the `new` constructor creates
and seeds the necessary RNGs. `new_using` is added for uses where
control over the RNGs is necessary.