Reseed RNGs when SGD is cloned
For parallel optimization, we clone the SGD struct. Since this clones
the internal state of the associated negative sampling RNG, all
threads will produce the same negative samples.

This change introduces a wrapper struct ReseedOnCloneRng. This struct
implements RngCore and delegates all RngCore methods to the wrapped
RNG. The Clone implementation of ReseedOnCloneRng does not clone the
inner RNG verbatim, but constructs a new instance of the RNG with
fresh entropy. This avoids correlation between the clones and the
cloned RNG.

ReseedOnCloneRng is now uses in SGD to ensure that clones of SGD have
different RNGs.