Introduce SimpleVocab struct and Vocab trait.
This commit adds the SimpleVocab type which is generic over the
stored types. The SimpleVocab can be used as an input and an ouput
lookup. To facilitate, this some of the lookup methods required by
vocabularies got delegated to a Vocab trait.
Construction of the SimpleVocab and SubwordVocab is done through
From<VocabBuilder> implementations.