Migrate to Rust 2018
We were still implicitly on Rust 2015, while at the same time using
post-2018 features. So:

- Switch to Rust 2018
- Also do CI builds on Rust 1.31

The motivation to also test builds against 1.31 is that this is the
first Rust version that supports 2018. Later versions have changes
that are not strictly Rust 2018. I want to stick to 1.31/2018 as the
lowest common denominator for a while, since Linux distributions are
sometimes trailing in Rust versions.